+++ shcDecoder/src/main/java/ca/bc/gov/shcdecoder/cache/FileManager.kt
@@ -2,6 +2,7 @@ package ca.bc.gov.shcdecoder.cache
 
 import ca.bc.gov.shcdecoder.model.Issuer
 import ca.bc.gov.shcdecoder.model.JwksKey
+import ca.bc.gov.shcdecoder.model.RevocationsResponse
 import ca.bc.gov.shcdecoder.model.Rule
 import java.util.Date
 
@@ -15,9 +16,7 @@ interface FileManager {
 
     suspend fun getRule(url: String): List<Rule>
 
-    suspend fun getRevocations(url: String): List<Pair<String, Date?>>
-
-    suspend fun getRevocationsCtr(url: String): Long?
+    suspend fun getRevocations(url: String): RevocationsResponse?
 
     suspend fun exists(url: String): Boolean
 }
